Renamed frontend page files and exported components
