+++ gui/src/pages/LogsPage.scss
