+++ gui/src/pages/RecipeDetailsPage.scss
