+++ gui/src/pages/SettingsPage.scss
